requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
langchain==0.3.24
langchain-community==0.3.22
langchain-chroma==0.2.3
langchain-huggingface==0.1.2
langchain-google-genai==2.1.3
huggingface-hub==0.30.2
sentence-transformers==4.1.0
pymongo==4.12.0
streamlit==1.44.1
requests==2.32.3
python-dotenv==1.1.0
google-genai==1.12.1
google-ai-generativelanguage==0.6.17